chore(cleanup): add account-related pages and refactor import paths
