+++ src/gui/plugins/menu_extension/plugin_homepage_menu_account.py
@@ -1,6 +1,6 @@
 import tkinter as tk
-from gui.accountpage.accountpage import AccountPage
-from gui.accountpage.reorder_account_widgets_page import (
+from gui.pages.account.accountpage import AccountPage
+from gui.pages.account.reorder_account_widgets_page import (
     ReorderAccountWidgetsWindow,
 )
 from gui.app.basewindow import BaseWindow
